BJS2-40691 Любой проект может выставлять вакансии о поиске специалистов. В будущем мы сможем расширить это в post_service — можно будет писать посты о вакансиях, нотификации прикрутим и т.д. Пока что просто создаём вакансии: можно выпустить несколько одинаковым вакансий, можно описывать в них требуемый опыт и т.д. Также на проекте обязательно должен быть человек, ответственный за вакансию.

+++ src/main/java/faang/school/projectservice/service/CandidateService.java
@@ -18,11 +18,11 @@ public class CandidateService {
         if (candidate == null) {
             log.info("Candidate with id {} did n't find", idCandidate);
         } else {
-            if (candidate.getCandidateStatus() == null || candidate.getCandidateStatus() != candidateStatusNotDelete) {
+            if (candidate.getCandidateStatus() != null
+                    && candidate.getCandidateStatus().name().equals(candidateStatusNotDelete.name())) {
                 candidateRepository.deleteById(idCandidate);
                 log.info("CandidateService. Candidate with id {} deleted", idCandidate);
             }
-
         }
     }
 }
